requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
deepface>=0.0.79
tf-keras>=2.15.0
tensorflow>=2.15.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
pillow>=10.0.0
fer>=22.5.1
spotipy>=2.23.0
requests>=2.31.0
